+++ pkg/ent/team/where.go
@@ -54,7 +54,7 @@ func IDLTE(id int) predicate.Team {
 }
 
 // Number applies equality check predicate on the "number" field. It's identical to NumberEQ.
-func Number(v int8) predicate.Team {
+func Number(v int) predicate.Team {
 	return predicate.Team(sql.FieldEQ(FieldNumber, v))
 }
 
@@ -69,42 +69,42 @@ func Password(v string) predicate.Team {
 }
 
 // NumberEQ applies the EQ predicate on the "number" field.
-func NumberEQ(v int8) predicate.Team {
+func NumberEQ(v int) predicate.Team {
 	return predicate.Team(sql.FieldEQ(FieldNumber, v))
 }
 
 // NumberNEQ applies the NEQ predicate on the "number" field.
-func NumberNEQ(v int8) predicate.Team {
+func NumberNEQ(v int) predicate.Team {
 	return predicate.Team(sql.FieldNEQ(FieldNumber, v))
 }
 
 // NumberIn applies the In predicate on the "number" field.
-func NumberIn(vs ...int8) predicate.Team {
+func NumberIn(vs ...int) predicate.Team {
 	return predicate.Team(sql.FieldIn(FieldNumber, vs...))
 }
 
 // NumberNotIn applies the NotIn predicate on the "number" field.
-func NumberNotIn(vs ...int8) predicate.Team {
+func NumberNotIn(vs ...int) predicate.Team {
 	return predicate.Team(sql.FieldNotIn(FieldNumber, vs...))
 }
 
 // NumberGT applies the GT predicate on the "number" field.
-func NumberGT(v int8) predicate.Team {
+func NumberGT(v int) predicate.Team {
 	return predicate.Team(sql.FieldGT(FieldNumber, v))
 }
 
 // NumberGTE applies the GTE predicate on the "number" field.
-func NumberGTE(v int8) predicate.Team {
+func NumberGTE(v int) predicate.Team {
 	return predicate.Team(sql.FieldGTE(FieldNumber, v))
 }
 
 // NumberLT applies the LT predicate on the "number" field.
-func NumberLT(v int8) predicate.Team {
+func NumberLT(v int) predicate.Team {
 	return predicate.Team(sql.FieldLT(FieldNumber, v))
 }
 
 // NumberLTE applies the LTE predicate on the "number" field.
-func NumberLTE(v int8) predicate.Team {
+func NumberLTE(v int) predicate.Team {
 	return predicate.Team(sql.FieldLTE(FieldNumber, v))
 }
